Support for the callback-based async IO jobs
Another type of the AsyncIO methods are now supported - in addition to
the legacy ones where the fiber is put to sleep inside AsyncIO,
"callback" based types of requests are added: the user will pass:

- SuspendableRequstHandler which will be used to notify the user
  that the operation has been completed
- Callback delegate to which the result of the actual IO operations are
  passed.

User can now sleep, and the callback will be called, updating the user's
state with the results of the aio operation, and then the user will be
called.

This allows for more advanced patterns where the user can choose where
to sleep, and where it can cancel the scheduled operations in case it
has been resumed by other means than AsyncIO system.