Add DelegateJobNotification implementation
DelegateJobNotification allows for the separation between
the policy how the particular suspendable job is
being suspended and how it should be resumed. It's moving this
concern from the AsyncIO, so every user of DelegateJobNotification
is due to implement it's own strategy providing suspend/resume
delegates.

The use case is of multiple-resume-points is also supported, by
providing no suspend delegate. In that case, context is forced
to suspend itself, in a place it finds fit.