Add Promise/Future struct definition
Promise/Future are used to transfer data during non-blocking async IO
data read from the AsyncIO (the producer) to the consumers (such as
request implementation). Promise and Future are two sides of the same coin -
the producer owns the promise instances from which the future structure
(containing _future_ result of the computation) is given to the consumer.