Allow multiple finish callback delegates for the AsyncIO job.
If there are multiple consumers in the chain consuming a single
nonblocking async IO read, we need to allow all of them to update state.
Instead of passing n number of delegates as a parameters to pread, pread
was changed to return the pointer to the Job, so `registerCallback`
method can be used.