Add Build.mak options to support more compiler and distro diversity
This patch adds a couple of little helpful build variables that may be
of assistance when building on more recent distros, or with compilers
other than `dmd-transitional`.

Ubuntu 18.04 (among other distros) expects position-independent code.
The new `USE_PIC` build variable allows the user to request this if the
compiler does not use the `-fPIC` flag by default (as is e.g. the case
with `dmd-transitional`).

LDC appears to be a bit more strict than DMD when it comes to generating
warnings.  The `ALLOW_WARNINGS` variable allow the user (or CI) to be a
bit more forgiving of problems that may arise in library space and are
therefore outside the app developer's immediate ability to correct.